requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
PyYAML>=6.0

# Machine Learning
lightgbm>=3.3.0
torch>=1.13.0
joblib>=1.2.0

# Time series and forecasting
meteostat>=1.6.0

# API and web
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.13.0

# Streamlit
streamlit>=1.20.0

# Utilities
requests>=2.28.0
python-dateutil>=2.8.0
tqdm>=4.64.0

# Development and testing (optional)
pytest>=7.2.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.20.0